Simplifying the sqlite binary management + Binary name uniformisation
