+++ packages/optimise-sqlite/bin/.gitignore
