+++ packages/optimise-sqlite/package.json
@@ -26,6 +26,7 @@
     "dependencies": {},
     "devDependencies": {
         "decompress": "4.2.0",
-        "fs-extra": "7.0.1"
+        "fs-extra": "7.0.1",
+        "tape": "4.9.2"
     }
 }
